+++ chaostoolkit/cli.py
@@ -6,12 +6,14 @@ import json
 import logging
 import os
 import sys
-from typing import List
+from typing import List, NoReturn
 import uuid
 
+from chaoslib import __version__ as chaoslib_version
 from chaoslib.exceptions import ChaosException, DiscoveryFailed, InvalidSource
 from chaoslib.discovery import discover as disco
 from chaoslib.experiment import ensure_experiment_is_valid, run_experiment
+from chaoslib.info import list_extensions
 from chaoslib.loader import load_experiment
 from chaoslib.notification import notify, DiscoverFlowEvent, InitFlowEvent, \
     RunFlowEvent, ValidateFlowEvent
@@ -181,6 +183,53 @@ def validate(ctx: click.Context, path: str) -> Experiment:
         sys.exit(1)
 
     return experiment
+
+
+@cli.command()
+@click.argument('target')
+@click.pass_context
+def info(ctx: click.Context, target: str) -> NoReturn:
+    """Display information about the Chaos Toolkit environment.
+
+    Available targets are:
+
+    * core: display the information about your version of the Chaos Toolkit
+
+    * extensions: display the list of installed extensions and plugins
+
+    * settings: display your current full settings
+    """
+    if target not in ["core", "settings", "extensions"]:
+        raise click.BadArgumentUsage("Invalid target")
+
+    if target == "core":
+        fmt = "{:<20}{:<10}"
+        click.secho(
+            fmt.format("NAME", "VERSION"),
+            fg='bright_blue')
+        click.echo(fmt.format("CLI", __version__))
+        click.echo(fmt.format("Core library", chaoslib_version))
+    elif target == "extensions":
+        fmt = "{:<40}{:<10}{:30}{:50}"
+        click.secho(
+            fmt.format("NAME", "VERSION", "LICENSE", "DESCRIPTION"),
+            fg='bright_blue')
+        extensions = list_extensions()
+        for extension in extensions:
+            summary = extension.summary.replace(
+                "Chaos Toolkit Extension for ", "")[:50]
+            click.echo(
+                fmt.format(
+                    extension.name, extension.version, extension.license,
+                    summary))
+    elif target == "settings":
+        settings_path = ctx.obj["settings_path"]
+        if not os.path.isfile(settings_path):
+            click.echo("No settings file found at {}".format(settings_path))
+            return
+
+        with open(settings_path) as f:
+            click.echo(f.read())
 
 
 @cli.command()