+++ CHANGELOG.md
@@ -4,10 +4,19 @@
 
 [Unreleased]: https://github.com/chaostoolkit/chaostoolkit/compare/0.16.0...HEAD
 
+### Added
+
+- add `info` command to display basic information such as version of the
+  toolkit core library or installed extensions. Display also the current
+  settings [#64][64]
+
+[64]: https://github.com/chaostoolkit/chaostoolkit/issues/64
+
 ### Changed
 
 - strip command name before sending it to check newer version as sometimes
   we get a tabulation character in there
+- swap `logger.warn` for `logger.warning` as the former is obsolete
 
 ## [0.16.0][] - 2018-09-19